requirements.txt
click==8.3.1
colorama==0.4.6
keyboard==0.13.5
numpy==2.4.1
pygame-ce==2.5.6
